+++ app/presenters/activity_presenter.rb
@@ -316,6 +316,12 @@ class ActivityPresenter < SimpleDelegator
     ActionController::Base.helpers.number_to_currency(super, unit: "£")
   end
 
+  def commitment
+    return if super.nil?
+
+    CommitmentPresenter.new(super)
+  end
+
   def linkable_activity_select_label
     "#{roda_identifier} (#{title})"
   end