+++ CHANGELOG.md
@@ -1224,6 +1224,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 ## [unreleased]
 
 - Update Commitment-importing script to infer `transaction_date` value from Activity values
+- Collect original commitment figure in the new activity form
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-131...HEAD
 [release-131]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-130...release-131